requirements.txt
# step1: python -m venv cuda
# step2: ./cuda/Scripts/Activate 
# step: pip install -r requirements.txt

# Base ----------------------------------------
PyJWT
requests
livekit-api
livekit-agents[google,images]~=1.0,>=1.0.18
livekit-plugins-noise-cancellation~=0.2
python-dotenv
asyncio